Replace certificate with a self-signed version
This commit updates the certificate to use SHA256 as the signature
algorithm so OpenSSL 1.1.1 doesn't complain with the "ca md too weak"
error.

I used the following commands to update the certificate:

$ openssl req -x509 -newkey rsa:4096 -sha256 -days 3650 -nodes \
   -keyout server.key -out server.crt \
   -subj "/CN=127.0.0.1/O=pytest-localserver/OU=Testing Dept." \
   -addext "subjectAltName=DNS:localhost"
$ cat server.key server.crt >server.pem

Note that with this setup, it's not possible to install a CA certificate
on the client so it can verify the server certificate. I'm not concerned
about that for now; this is just a proof of concept that an updated
certificate makes the plugin work.