Update documentation of SecureContentServer with new openssl commands
This commit changes the docstring for SecureContentServer to describe
how to create a self-signed certificate or a certificate signed by
a private CA using OpenSSL 1.1.1. In particular, the -sha256 option is
necessary to prevent "ca md too weak" errors with that version of
OpenSSL (which is the minimum one required by Python 3.10).