Add pyproject.toml
This commit adds a pyproject.toml file specifiny the basic requirements
to use setuptools as the build backend. It uses isolated builds, as
necessary to make tox use the configuration in pyproject.toml when
building the package to use for testing. Otherwise it would just run
setup.py, which might miss the configuration in pyproject.toml.