Add setuptools_scm as a dependency
The version requirements were determined by testing with various old
versions of setuptools and setuptools_scm. Using setuptools_scm <3.4.1
causes the build to fail, and using setuptools <42 means it fails to
automatically determine the version number. Of course, in practice I
expect that people will generally be using relatively recent versions of
both packages, but the listed requirements express the minimum versions
that appear to be compatible with this process.