Remove explicit version specification
This commit removes the explicit listing of the version number in both
__init__.py and setup.py, and replaces the former with the version
number computed by setuptools_scm. As a prerequisite to that, I've added
a configuration item in pyproject.toml to have setuptools_scm write
the version number to a file.

I've imported the version number under the name `VERSION` to keep it
compatibile with previous versions of pytest_localserver. If `version`
is a more standard identifier, we might want to change or add that in
the future.