Disable fail-fast in Github Actions configuration
By default, Github Actions has the fail-fast option enabled, which means
that the first job to fail in a matrix causes the whole matrix to be
immediately canceled. This makes it impossible to identify when a bug
causing a job failure is specific to certain versions of Python. In this
commit I'm setting the option to false so that each job will run to
completion even if one fails.