Implement optional chunk encoding in ContentServer
This commit adds an implementation of chunk encoding to ContentServer.
The decision of whether to apply chunking is controlled by an enum,
which takes values YES (apply chunking), NO (don't apply chunking), or
AUTO (apply chunking only if requested by the Transfer-Encoding header).

I set the default value of the flag to NO for backwards compatibility,
but it could be changed to AUTO in the future, which would probably make
more sense.