Limit addr to (host, port) even for IPv6
The address of an IPv6 socket as returned by getsockname() is
a four-element tuple, rather than a two-element tuple as expected by
the existing pytest_localserver.smtp module code. In this commit I'm
extracting the first two elements of the tuple so that a Server's addr
attribute will be a 2-tuple of (host, port) regardless of whether
the underlying socket is an IPv4 or IPv6 socket.