Make aiosmtpd dependency optional
Move the aiosmtpd dependency to "smtp" extra, and handle the lack of it
gracefully in tests.  The aiosmtpd package is no longer maintained
and had major bugs that lead to it being queued for removal from Gentoo.
FWICS all packages in Gentoo that are using pytest-localserver are only
using the HTTP support.