Bump pytest max version to 8
We have a constraint that limits the maximum supported version of pytest
to the current major version, under the assumption that a new major
version release may break backwards compatibility. Now that pytest 7 has
been tested and is known to work with pytest-localserver, we can relax
the constraint to mark it as supported.