Use older Ubuntu for testing with Python 3.5 and 3.6
This commit changes the Github Actions workflow configuration to use
an older version of Ubuntu for testing on Python 3.5 and 3.6 - namely
Ubuntu 20.04 instead of the current default 22.04. The runners that
Github provides with Ubuntu 22.04 don't support versions of Python
prior to 3.7.

We will remove support for Python <3.7 at some point, when a code change
makes it necessary, but for now there happens to be nothing in the code
that breaks with Python 3.5 or 3.6 and it's easy enough to support those
versions, which is why I'm not removing support for them entirely yet.

I also upgraded the workflow configuration from the beta version of
Python 3.11 to the final release version.