Add a GitHub workflow for pushing to PyPI and Test PyPI
In this commit I'm adding a new workflow which will build distributable
packages for pytest-localserver and upload them to PyPI and Test PyPI.
The uploads are treated as deployments, so they will not run until
someone manually approves them.

A future enhancement I'd like to make is to actually run tests using
the same package that is going to be uploaded. That's going to involve
tweaking tox to install from a prebuilt package rather than directly
from the source code, so it will take some time to figure that out.