Update the AUTHORS file
We haven't been updating the AUTHORS file as various people have added
their contributions to the project, so in this commit I'm catching up on
those updates. It's important for both legal and ethical reasons that
everyone gets appropriate credit for their contributions.

I've taken the list of authors from the commit history. The list of
authors in the file is ordered chronologically by their first commit or
first interaction with the project (as far as I'm aware), which seems
consistent with the order that was there before.