Remove obsolete test code in runtests.py
The script runtests.py contains an archive of a very old version of
pytest along with a custom mechanism to import it. That used to be
the standard method of running tests (so I hear, anyway), but now it's
no longer relevant because we use tox for dependency management in
testing. So I'm removing the script. Having a large opaque blob of data
is not the greatest thing :-)