Make pypi deployment dependent on test-pypi deployment
This commit changes the deployment to PyPI to be dependent on successful
deployment to Test PyPI. Without this we would have to manually check
that the deployment to Test PyPI succeeded before deploying to PyPI, and
there's really no point in making that a manual process. (Or, a person
could just deploy to PyPI directly without even waiting for the Test
PyPI deployment to succeed and in that case there's no point in even
deploying to Test PyPI first.)

I was inspired to do this by looking at the release workflow for some
other project, but I forget which one it was.