Move SMTP driver code into a function
pytest_localserver.smtp allows running itself as a module to start up
an SMTP server, which is implemented by a bit of driver code guarded by
an `if __name__ == "__main__"` block. Typical best practice is to put
that code in a `main()` function, which is what I'm doing in this
commit.