Update README not to specify the wrong type for the SMTP server
The README file had said that the SMTP server would be an instance of
smtpd.SMTPServer, but that's actually no longer true now that we don't
use the smtpd module anymore. In this commit I'm updating that bit of
documentation not to specifically reference smtpd.SMTPServer. Instead,
it's a bit vague about what kind of SMTP server is provided, only
promising an API "similar to smtpd.SMTPServer". This gives us freedom
to change the exact API offered by that class if necessary.