Modify a docstring to avoid exceeding line length limit
A long line in this docstring was causing the flake8 pre-commit check to
fail, so I edited the comment to be compliant. I shortened the lines to
about 80 characters for consistency with other comments, even though
that's more of a change than is strictly required to pass flake8. With
this change, we can start requiring pre-commit to pass for all merges.

I also changed a note to use Restructured Text syntax, for compatibility
with other docstrings in the project.