Bump required pytest version to 4.6
We have started getting an error with pytest>=4,<4.6 when setuptools>=71
(or typeguard>=4.2) is also installed, due to a conflict with older
pytest and setuptools' vendored version of typeguard. Out of various
ways to resolve this conflict, this is the only one that doesn't involve
putting an upper bound on the version of some package.

It should still be technically possible to use older versions of pytest
along with setuptools<71, if someone really needs to do that, but there
isn't any way to express that dependency constraint, and I'm not going
to go to any trouble to support long-obsolete versions of packages.