Remove driver code
Since this package is meant to be installed and imported, it doesn't
really make sense anymore to have code in the modules that allows those
modules to be executed as individual Python scripts. So I'm removing
that code in this commit.

I considered adding tests to replicate the functionality of the code
being removed, but there didn't seem to be much point. The removed code
just does a basic check that the server can start and stop, and relies
on some external client to send HTTP or SMTP requests, and our existing
tests cover all that functionality already.