Add Python 3.14 testing to CI and refactor version matrix
The main purpose of this commit is to enable testing on Python 3.14 in
CI. But Python 3.14 requires a different minimum version of pytest from
any earlier version, meaning that we now have three different minimum
pytest versions that need to be tested depending on the Python version,
and that's very complicated to express by including and excluding
specific configurations as we were doing before. So I refactored
the matrix of versions to take advantage of the ability to have objects
as matrix values, meaning that each Python version can be accompanied by
the specific minimum pytest version it requires in the matrix. This is
much easier to understand and maintain, and it should also have us
running fewer CI jobs overall (since we're no longer going to run
combinations like pytest 6.2.4 on Python 3.7-3.9, which served no
purpose).