Combine CI jobs into one per Python version
We run a lot of tests in CI for such a small project, and in the past
each tox environment has been run in a separate GitHub Actions job,
but we don't really need to do that. The reason we've been splitting
them up that way is to make it clear at a glance what configurations are
failing when a CI run fails, but in practice we've never really had
a failing workflow where we needed to look at all the different failing
tox runs to understand the problem; looking at one failure is generally
enough.

So, in this commit, I'm consolidating the GHA jobs into one per Python
version which will run all the tox environments for that Python version.
This cuts down a bit on setup work, and might reduce the load on PyPI
a bit, since downloaded packages can potentially be cached across
different tox runs for the same Python version.