Add a summary job that checks whether all tests passed
We can use this to add a status check to ensure that only PRs passing
all tests can be merged.

The commit ID I used for the alls-green action is taken from the current
tip of the release/v1 branch. (It's not clear whether the action is
maintained, but using a commit hash means we don't have to care about
any future changes unless we actively decide to.)