Update the AUTHORS file with recent contributors
We haven't consistently been updating the AUTHORS file as various people
have added their contributions to the project (again, my bad, code
review should have caught that and we really should add an automated
check for it at some point), so in this commit I'm catching up on those
updates. It's important for both legal and ethical reasons that everyone
gets appropriate credit for their contributions.

I've taken the list of authors from the commit history. The list of
authors in the file is ordered chronologically by their first commit or
first interaction with the project (as far as I'm aware).